Change un type de place
